edit denomination
