+++ README.md
@@ -1,24 +1,33 @@
 # kotlin-blackjack
 
 
-### 1. 객체 정리
+### 1. 객체 정리와 역할일, 해야 할 
 
 #### Deck
-- 기본 덱은 52장으로 이루어져 있다
-- 덱을 섞을 수 있다
-- 카드 한 장을 뽑을 수 있다
+- 기본 덱은 52장으로 이루어져 있다(특징)
+- 덱을 섞을 수 있다 (일)
+- 카드 한 장을 뽑을 수 있다 (일)
 
 #### Card
-- 카드는 슈트와 끗수로 이루어져 있다
-- 슈트와 끗수가 동일한 경우 같은 카드이다.
+- 카드는 슈트와 끗수로 이루어져 있다 (특징)
+- 슈트와 끗수가 동일한 경우 같은 카드이다. (특징 & 일)
 
 #### Denomination
-- 숫자는 그대로 계산한다
-- J, Q, K는 10점으로 계산한다
+- 숫자는 그대로 계산한다 (특징)
+- J, Q, K는 10점으로 계산한다 ()
 - 에이스를 포함한 카드의 합이 21이 넘는 경우 에이스는 1로 계산한다
 - 에이스를 포함한 카드의 합이 21을 넘지 않는 경우 에이스는 11로 계산한다
+
 #### Suit
 - 모양별로 4종류의 슈트가 있다 (CLUBS, DIAMONDS, HEARTS, SPADES)
+
+#### Player
+- 처음에 2장의 카드를 받는다
+- 그리고 한 장의 카드를 추가적으로 받을지 말듯 결정한다.(yes or no)
+
+#### 결과
+- 각 참가자의 모든 카드를 합친다.
+
 
 ### 2. step2에 요구사항
 - 게임에 참여할 이름 입력
@@ -26,5 +35,6 @@
 - 첫번째 플레이어가 아니오 할때까지 카드 추가 여부 문의
 - 두번쨰 플레이어에거 첫번째 플레이어에게 했던 문의를 한다.
 - 결과를 출력한다.
+
