--- src/main/kotlin/blackjack/Card.kt
@@ -1,3 +1,0 @@
-package blackjack
-
-data class Card(var Denomination: Denomination, var suit: Suit)
